Create #5_Smallest_Multiple
Problem: 2520 is the smallest number that can be divided by each of the numbers from 1 to 10 without any remainder.

What is the smallest positive number that is evenly divisible by all of the numbers from 1 to 20?

This program takes about 5.5 seconds to complete on average.

I didn't include some numbers in my list because they were unnecessary. Example: Anything that is divisible by 18 will also be divisible by 2, 3, 6 and 9. Approaching the problem like this allowed me to reduce the number of iterations necessary. I also made the first three numbers in my list all odd for the same effect; less iterations.