Create #7 10001st Prime_Old&Slow
Started out by using this code to get the 10001st prime. Each consecutive new prime found took an exponentially longer time to find. By the time I got to the 1200th prime it was taking around a minute to find the next 100 primes. at 1500 it was taking aprox. 5 minutes to get the next 100 primes